Moved _setup_recording to KilosortBase
